Updated Global State Variables
