+++ .gitignore
@@ -24,3 +24,6 @@ yarn-error.log*
 
 # env files
 *.env
+
+# logs
+/logs
