+++ fele-client-service/client-api/scripts/network.js
@@ -1,6 +1,7 @@
 const { createDatabase, deleteDatabase, insertToDatabase } = require('../../utils/db')
 const path = require("path");
 const fs = require('fs');
+const { GLOBAL_STATE } = require('../../utils/constants');
 
 var dbPrefix = "fele__"
 var chaincodeDirectory = "../../../chaincode/"
@@ -24,12 +25,14 @@ const createNetwork = async(networkConfigJSON, networkName) => {
 }
 
 const useNetwork = (username, localOrg, networkName) => {
-  const { localUsers, felenetworks } = localOrg
+  console.log(GLOBAL_STATE);
+  const { localUsers, felenetworks } = GLOBAL_STATE.localOrg
   const feleUser = {}
   const localUserIndex = localUsers.findIndex((localUser) => (localUser.username === username));
   if(localUserIndex != -1) {
         const felenetworkIndex = felenetworks.findIndex((felenetwork) => felenetwork.felenetId === networkName);
         if(felenetworkIndex != -1) {
+            feleUser.network = felenetworks[felenetworkIndex];
             const isMappingPresentForUser = felenetworks[felenetworkIndex].mappings.findIndex((mapping) => mapping.from === username);
             if(isMappingPresentForUser != -1) {
                 feleUser.user = felenetworks[felenetworkIndex].mappings[isMappingPresentForUser].to
