+++ fele-client-service/utils/constants.js
@@ -1,7 +1,14 @@
 const DB_PREFIX = "fele__"
 const CHANNEL_ID_PREFIX = "channel-"
+const GLOBAL_STATE = {
+    network: "",
+    localOrg: {},
+    localUser: {},
+    feleUser: {}
+};
 
 module.exports = {
     DB_PREFIX,
-    CHANNEL_ID_PREFIX
+    CHANNEL_ID_PREFIX,
+    GLOBAL_STATE
 }
