+++ logs/logs.log
@@ -1,0 +1,12 @@
+error: Label: Feb-26-2023 13:09:54: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:12:34: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:13:12: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:15:52: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:16:26: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:18:48: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:30:52: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 13:31:33: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 14:09:40: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 14:12:08: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 14:37:58: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
+error: Label: Feb-26-2023 14:39:18: Error retrieving document from database:  Unexpected status code while fetching documents from the database: 401
