+++ package.json
@@ -24,7 +24,7 @@
     "express": "^4.18.2",
     "fs-extra": "^11.1.0",
     "g": "^2.0.1",
-    "jest": "^29.4.1",
+    "jest": "^29.4.3",
     "node-couchdb": "^2.0.0",
     "swagger-ui-express": "^4.6.0",
     "winston": "^3.8.2"
